'chore: saved into temp branch via 4got.save2branch'

+++ bin/git/branch.prune/branch.prune.sh
@@ -1,9 +1,19 @@
+#!/usr/bin/env bash
 
-# TODO: add args
-# TODO: add error codes
-# TODO: ask before deleting
+# 4got.branch.prune
+#
+# @arg $1 string git branch name
+# @example
+#     4got.basename develop
+#
+# @exitcode 0 If successful
+# @exitcode 1 If an empty string passed to $1
 function 4got.branch.prune() {
-  # TODO: add condition for not passed $1
+  if [[ -z "$1" ]]; then
+    echo "Usage: $0 string [branch_name]"
+    return 1
+  fi
+  # TODO: ask before deleting
   git branch -D "$1"
   git push origin --delete "$1"
 }
